+++ src/main/java/dto/TransactionCreateDtoIn.java
@@ -1,5 +1,5 @@
 package dto;
 
-public record TransactionCreateDtoIn(int userId, int categoryId, Integer transactionAmount, String transactionDate, String transactionType, String transactionDescription, String transactionFrequency) {
+public record TransactionCreateDtoIn(int userId, int categoryId, Double transactionAmount, String transactionDate, String transactionType, String transactionDescription, String transactionFrequency) {
 
 }
